Updated pycaret regression models to begin to shape the inference dataframe
I am in the process of setting up the inference dataframe which will be used to store the predicted values of traffic_abs and traffic_norm.

I am currently attempting to pull values from columns in the cleaned visitor center data frame and merge them into corresponding columns in the inference dataframe, matching on date/time at the hourly level.